Refactored Nav auth sections into own components
