+++ client/src/components/Navbar/Navbar.js
@@ -3,19 +3,18 @@ import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
-import { Grid, Avatar, Button } from '@material-ui/core';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import { NavLink, useHistory } from 'react-router-dom';
-import isEmpty from 'is-empty';
-import { disconnectClient } from '../../socketio-client';
+import { Grid } from '@material-ui/core';
+import { NavLink } from 'react-router-dom';
 
 // ASSETS
 import logo from '../../assets/images/logos/logo.png';
 
 // CONTEXT
 import { userState } from '../../provider/UserContext';
-import { LOGOUT_USER } from '../../provider/constants';
+
+// COMPONENTS
+import AuthenticatedNav from './AuthenticatedNav';
+import UnauthenticatedNav from './UnauthenticatedNav';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -34,22 +33,6 @@ const useStyles = makeStyles((theme) => ({
     marginLeft: 10,
     marginRight: 10,
   },
-  link: {
-    textDecoration: 'none',
-    fontSize: 12,
-    color: 'black',
-    margin: '0 1rem',
-    textTransform: 'uppercase',
-    fontWeight: 500,
-    '&.active': {
-      borderBottom: '2px solid',
-      borderBottomColor: '#69E781',
-    },
-    '&:hover': {
-      color: 'grey',
-      background: 'transparent',
-    },
-  },
   toolbar: {
     display: 'flex',
     zIndex: 10,
@@ -59,81 +42,9 @@ const useStyles = makeStyles((theme) => ({
 
 const Navbar = () => {
   const classes = useStyles();
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const history = useHistory();
   const {
-    state: { isAuthenticated, user },
-    dispatch,
+    state: { isAuthenticated },
   } = useContext(userState);
-
-  const handleClick = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const logout = () => {
-    dispatch({ type: LOGOUT_USER });
-    fetch('/logout', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-    });
-    disconnectClient();
-    history.push('/login');
-    handleClose();
-  };
-
-  const renderNavbarLoggedUser = () => (
-    <div>
-      <NavLink color="inherit" to="/explore" className={classes.link}>
-        Explore
-      </NavLink>
-      <NavLink color="inherit" to="/launch" className={classes.link}>
-        Launch
-      </NavLink>
-      <Button
-        aria-controls="profile"
-        aria-haspopup="true"
-        onClick={handleClick}
-      >
-        <Avatar alt="User" src={!isEmpty(user.avatar) ? user.avatar : null}>
-          {user.name[0]}
-        </Avatar>
-      </Button>
-      <Menu
-        id="profile-menu"
-        anchorEl={anchorEl}
-        keepMounted
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-      >
-        <MenuItem
-          component={NavLink}
-          to={`/users/${user._id}`}
-          onClick={handleClose}
-        >
-          Profile
-        </MenuItem>
-        <MenuItem onClick={logout}>Logout</MenuItem>
-      </Menu>
-    </div>
-  );
-
-  const renderNavbarNotLoggedUser = () => (
-    <div>
-      <NavLink color="inherit" to="/explore" className={classes.link}>
-        Explore
-      </NavLink>
-      <NavLink color="inherit" to="/login" className={classes.link}>
-        Login
-      </NavLink>
-      <NavLink color="inherit" to="/signup" className={classes.link}>
-        Sign Up
-      </NavLink>
-    </div>
-  );
 
   return (
     <div className={classes.root}>
@@ -144,9 +55,11 @@ const Navbar = () => {
             <Typography variant="h6">Product Launch</Typography>
           </Grid>
 
-          {isAuthenticated
-            ? renderNavbarLoggedUser()
-            : renderNavbarNotLoggedUser()}
+          {isAuthenticated ? (
+            <AuthenticatedNav />
+          ) : (
+            <UnauthenticatedNav />
+          )}
         </Toolbar>
       </AppBar>
     </div>
